requirements.txt
langchain-core
langchain-community
chromadb
sentence-transformers
python-dotenv
groq
